Various Updates
Stopped the counting of attempts and deaths when replaying breakthroughs

Reimplemented the feature of the program starting the next level with the winning genome of the last level. This feature was broken when I implmented the replay of breakthroughs in this program. This reimplmentation restores this feature.

Changed the formatting of the information written to the file 'fitnesstracker.txt' to be more regular.