Refactored Bonus areas, and fitness function ranges
Refactored parts of the code pertaining to the special ranges for MarioBot where the fitness function to reward Mario is modified. This is refactored such that certain ranges are only activated for a specific ROM with an MD5 hash that derived from a .nes file that is in a directory called "LoadedROM" in the local directory.

Bonus areas are also refactored similarly where certain bonus areas are activated based upon what ROM is found (and the MD5 hash that is derived from it) in the "LoadedROM" directory.