Several Updates
1. Added castleVictoryTime variable to keep track of the time elapsed since Mario beat a castle level. After 30 seconds of beating a castle 
level, the program will force a replay of breakthroughs. This change is introduced because at the end of the game (8-4), the program has no 
way of replaying breakthroughs since the game will never be in a state of 'prelevel' as defined in this program/script.
2. Took the block of code that was responsible for triggering the replays and made it into it's own function. This helps reduce the amount of 
code that would be needed for the change in the previous point.
3. Fixed a bug in drawing the neural net where the program would, sometimes under unknown conditions or circumstances, have a neuron (
neuron.x) as nil. I first used a work around for this issue, but now the neuron will always have a value and not return nil.
4. At the beginning of a level, the program will write all GSIDs in the current generation to a file.