Fixed File Scanning
Made the code more adaptable to finding winning genomes for levels on start. The program now looks for folders within the 'backups' folder that match the World-Level scheme and loads the winning genome found within this folder. Before, the program would look through a fixed amount folders which is not ideal where World and Level counts/amounts could vary